🔨 drop vertical color legend observer
